OTP 19 compatibility.
OTP 19 has landed and has the changes required to do EAP-TLS.
Notes for folks who are using older OTP versions are in the README.

OTP 19 also performs TLS handshake record batching for us. So, we now
test to see if we're running with ssl application version >= 8.x and
-if we are- assume that the handshake records have been given to us in
a single batch, rather than one at a time.